+++ apps/frontend/src/app/1_atoms/LoaderWithLogo/LoaderWithLogo.tsx
@@ -2,7 +2,8 @@ import React, { FC } from 'react';
 
 import { t } from 'i18next';
 
-import { ReactComponent as LogoSVG } from '../../../assets/images/sovryn-small-logo.svg';
+import { Lottie } from '@sovryn/ui';
+
 import { translations } from '../../../locales/i18n';
 
 type LoaderWithLogoProps = {
@@ -11,7 +12,9 @@ type LoaderWithLogoProps = {
 
 export const LoaderWithLogo: FC<LoaderWithLogoProps> = ({ text }) => (
   <div className="flex flex-col justify-center items-center flex-grow">
-    <LogoSVG className="sv-logo" />
+    <div className="w-64">
+      <Lottie animation="loading" />
+    </div>
     <div className="sv-loading-text">
       {text ? text : t(translations.loader.loading)}
     </div>